# tipu-api/openapi.yaml
openapi: 3.0.0
info:
  title: TIPU Academy API
  version: 1.0.0
  description: Online tutoring platform API with Firebase and Stripe integration
  contact:
    name: TIPU Academy Team
    email: support@tipuacademy.com
  license:
    name: MIT
servers:
  - url: http://localhost:8888
    description: Development server
  - url: https://api.tipuacademy.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT authentication token
  schemas:
...
    Booking:
      type: object
      required:
        - id
        - studentId
        - tutorId
        - subject
        - level
        - scheduledAt
        - status
        - price
      properties:
        id:
          type: string
          description: Booking ID
          example: booking123
        studentId:
          type: string
          description: Student user ID
          example: student123
        tutorId:
          type: string
          description: Tutor user ID
          example: tutor123
        subject:
          type: string
          enum:
            - Maths
            - Physics
            - Computer Science
            - Python
          description: Subject being taught
          example: Maths
        level:
          type: string
          enum:
            - GCSE
            - A-Level
          description: Education level
          example: GCSE
        scheduledAt:
          type: string
          format: date-time
          description: Session date and time
          example: "2025-11-23T10:00:00Z"
        duration:
          type: number
          description: Session duration in minutes
          minimum: 15
          maximum: 180
          default: 60
          example: 60
        status:
          type: string
          enum:
            - pending
            - confirmed
            - completed
            - cancelled
            - declined
          description: Booking status
          example: confirmed
        price:
          type: number
          description: Price in pence (£45.00 = 4500)
          minimum: 0
          example: 4500
        isPaid:
          type: boolean
          description: Payment status
          example: true
        paymentIntentId:
          type: string
          description: Stripe Payment Intent ID
          example: pi_123456789
        meetingLink:
          type: string
          format: uri
          description: Zoom/Google Meet URL
          example: https://zoom.us/j/123456789
        recordingUrl:
          type: string
          format: uri
          description: Session recording URL
          example: https://storage.googleapis.com/recordings/session123.mp4
        lessonReport:
          type: object
          properties:
            topicsCovered:
              type: string
              description: Topics covered in the lesson
              example: Quadratic equations, factorization
            homework:
              type: string
              description: Homework assigned
              example: Complete exercises 1-10 on page 45
            notes:
              type: string
              description: Additional notes
              example: Student is progressing well
            completedAt:
              type: string
              format: date-time
              example: "2025-11-23T11:00:00Z"
        declineReason:
          type: string
          description: Reason for declining (if applicable)
          example: Schedule conflict
        createdAt:
          type: string
          format: date-time
          example: "2025-11-20T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-21T09:15:00Z"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Resource not found
        message:
          type: string
          description: Detailed error message
          example: The requested booking does not exist
        code:
          type: string
          description: Error code
          example: BOOKING_NOT_FOUND
paths:
...
  /v1/bookings:
    post:
      tags:
        - Bookings
      summary: Create new booking request
      description: Student creates a new tutoring session booking request. Tutor must accept before session is confirmed.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tutorId
                - subject
                - level
                - scheduledAt
                - price
              properties:
                tutorId:
                  type: string
                  description: ID of the tutor to book
                  example: tutor123
                subject:
                  type: string
                  enum:
                    - Maths
                    - Physics
                    - Computer Science
                    - Python
                  description: Subject to be taught
                  example: Maths
                level:
                  type: string
                  enum:
                    - GCSE
                    - A-Level
                  description: Education level
                  example: GCSE
                scheduledAt:
                  type: string
                  format: date-time
                  description: Requested session date and time
                  example: "2025-11-23T10:00:00Z"
                price:
                  type: number
                  description: Price in pence (£45.00 = 4500)
                  minimum: 0
                  example: 4500
                duration:
                  type: number
                  description: Session duration in minutes
                  minimum: 15
                  maximum: 180
                  default: 60
                  example: 60
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized - No valid JWT token provided
        "404":
          description: Tutor not found

    get:
      tags:
        - Bookings
      summary: Get user's bookings
      description: Retrieve bookings for the authenticated user. Students see their bookings, tutors see bookings assigned to them.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - pending
              - confirmed
              - completed
              - cancelled
              - declined
          description: Filter by booking status
          example: confirmed
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of bookings to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of bookings to skip (for pagination)
      responses:
        "200":
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
              example:
                bookings:
                  - id: booking123
                    studentId: student123
                    tutorId: tutor456
                    subject: Maths
                    level: GCSE
                    status: confirmed
                    price: 4500
                    scheduledAt: "2025-11-23T10:00:00Z"
        "401":
          description: Unauthorized
  /v1/bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      description: Retrieve detailed information about a specific booking
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Booking ID
          example: booking123
      responses:
        "200":
          description: Booking details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
        "404":
          description: Booking not found
  /v1/bookings/{id}/accept:
    post:
      tags:
        - Bookings
      summary: Accept booking request (Tutor only)
      description: Tutor accepts a booking request and provides the meeting link for the session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Booking ID to accept
          example: booking123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - meetingLink
              properties:
                meetingLink:
                  type: string
                  format: uri
                  description: Zoom, Google Meet, or other video call URL
                  example: https://zoom.us/j/123456789
      responses:
        "200":
          description: Booking accepted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Booking accepted successfully
        "400":
          description: Invalid meeting link or booking already accepted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not the assigned tutor
        "404":
          description: Booking not found
  /v1/bookings/{id}/decline:
    post:
      tags:
        - Bookings
      summary: Decline booking request (Tutor only)
      description: Tutor declines a booking request. The system will notify other tutors for the subject.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Booking ID to decline
          example: booking123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: Reason for declining the booking
                  minLength: 3
                  maxLength: 500
                  example: Schedule conflict - already booked at this time
      responses:
        "200":
          description: Booking declined successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Booking declined
        "400":
          description: Invalid reason or booking already declined
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not the assigned tutor
        "404":
          description: Booking not found
  /v1/bookings/{id}/lesson-report:
    post:
      tags:
        - Bookings
      summary: Submit lesson report (Tutor only)
      description: Tutor submits a detailed report after completing a session, including topics covered, homework, and notes
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Booking ID
          example: booking123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topicsCovered
              properties:
                topicsCovered:
                  type: string
                  description: Topics and concepts covered in the lesson
                  minLength: 10
                  maxLength: 2000
                  example: Quadratic equations, factorization methods, completing the square
                homework:
                  type: string
                  description: Homework assignments given to the student
                  maxLength: 1000
                  example: Complete exercises 1-10 on page 45, practice factoring problems
                notes:
                  type: string
                  description: Additional notes about student progress or concerns
                  maxLength: 1000
                  example: Student is progressing well, showed good understanding of the concepts
      responses:
        "200":
          description: Lesson report submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lesson report submitted successfully
        "400":
          description: Invalid report data or session not completed
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not the assigned tutor
        "404":
          description: Booking not found
...